Inline headers from field
